skip fuzzy entries
git-svn-id: https://svn.r-project.org/R/trunk@59191 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/array.c
@@ -125,13 +125,13 @@ SEXP attribute_hidden do_matrix(SEXP call, SEXP op, SEXP args, SEXP rho)
     {
         if (lendat > (double)nc * INT_MAX)
             error("data is too long");
-        nr = (int)ceil(lendat / (double)nc);
+        nr = (int)ceil((double)lendat / (double)nc);
     }
     else if (miss_nc)
     {
         if (lendat > (double)nr * INT_MAX)
             error("data is too long");
-        nc = (int)ceil(lendat / (double)nr);
+        nc = (int)ceil((double)lendat / (double)nr);
     }
 
     if (lendat > 0)
